♻️ refactor : 프론트·백엔드 성능 구조 개선: 게임/로비/전역채팅 분리, 커뮤니티 patch 렌더 도입, accounts/season 서비스 슬림화
