+++ templates/base.html
@@ -150,6 +150,7 @@
     <script src="{% static 'js/components/session-guard.js' %}"></script>
     <script src="{% static 'js/components/notifications.js' %}"></script>
     <script src="{% static 'js/components/party_invites.js' %}"></script>
+    <script src="{% static 'js/components/dom_patch_list.js' %}"></script>
     <script src="{% static 'js/components/global_dm_ui.js' %}"></script>
     <script>
         (function() {
@@ -271,6 +272,7 @@
             updateStatus();
         })();
     </script>
+    <script src="{% static 'js/components/global_dm_channels.js' %}"></script>
     <script src="{% static 'js/components/global_dm.js' %}"></script>
     {% block extra_js %}{% endblock %}
 </body>
