+++ templates/chess/game.html
@@ -3012,7 +3012,9 @@ body.competitive-room #rematch-modal {
 {% block extra_js %}
 <script src="{% static 'js/components/game_board_ui.js' %}"></script>
 <script src="{% static 'js/components/game_socket_client.js' %}"></script>
+<script src="{% static 'js/components/game_socket_handlers.js' %}"></script>
 <script src="{% static 'js/components/game_status_ui.js' %}"></script>
+<script src="{% static 'js/components/game_overlay_ui.js' %}"></script>
 <script src="{% static 'js/components/game_moves_cache.js' %}"></script>
 <script src="{% static 'js/components/game_analysis_ui.js' %}"></script>
 <script src="{% static 'js/components/game_side_panels_ui.js' %}"></script>
