+++ templates/chess/lobby.html
@@ -2147,5 +2147,6 @@
 <script src="{% static 'js/components/context-menu.js' %}"></script>
 <script src="{% static 'js/components/modal_launcher.js' %}"></script>
 <script src="{% static 'js/components/lobby_presence.js' %}"></script>
+<script src="{% static 'js/components/lobby_room_ui.js' %}"></script>
 <script src="{% static 'js/pages/lobby.js' %}"></script>
 {% endblock %}
